requirements.txt
python-pptx
Pillow
transformers
torch
easyocr
paddleocr
paddlepaddle
numpy
requests
